+++ server/components/native/middleware.js
@@ -108,6 +108,7 @@ export function getNativeFallbackMiddleware({
                 <!DOCTYPE html>
                 <head>
                     <meta name="viewport" content="width=device-width, initial-scale=1" />
+                    <link rel="manifest" href="/.well-known/manifest.webmanifest">
                     <title>Native Fallback</title>
                 </head>
                 <body data-nonce="${ cspNonce }" data-client-version="${ client.version }">
