+++ src/native/fallback/fallback.js
@@ -1,9 +1,12 @@
 /* @flow */
 
-import { cleanup } from 'belter/src';
-import type { ZalgoPromise } from 'zalgo-promise/src';
+import { cleanup, once, stringifyErrorMessage } from 'belter/src';
+import { ZalgoPromise } from 'zalgo-promise/src';
+import { FUNDING, FPTI_KEY } from '@paypal/sdk-constants/src';
 
-import { getPostRobot } from '../../lib';
+import { getLogger, getPostRobot } from '../../lib';
+import { FPTI_CUSTOM_KEY, FPTI_TRANSITION } from '../../constants';
+import {  isAppInstalled } from '../lib';
 
 import { MESSAGE } from './constants';
 
@@ -30,7 +33,30 @@ export function setupNativeFallback({ parentDomain = window.location.origin } : 
             .then(({ data }) => data);
     };
 
-    sendToParent(MESSAGE.DETECT_WEB_SWITCH);
+    const { env, fundingSource } = window.xprops;
+
+    const appInstalledPromise = once(isAppInstalled({ fundingSource, env })
+        .catch(err => {
+            getLogger().info('native_popup_android_app_installed_error')
+                .track({
+                    [FPTI_KEY.TRANSITION]:      FPTI_TRANSITION.NATIVE_POPUP_ANDROID_APP_ERROR,
+                    [FPTI_CUSTOM_KEY.ERR_DESC]: `Error: ${ stringifyErrorMessage(err) }`
+                }).flush();
+            
+            return ZalgoPromise.resolve(null);
+        }));
+
+    appInstalledPromise.then(app => {
+        if (fundingSource === FUNDING.VENMO && app.version && app.id) {
+            getLogger().info('native_fallback_retry_venmo_app_switch')
+                .track({
+                    [FPTI_KEY.TRANSITION]:      FPTI_TRANSITION.NATIVE_FALLBACK_RETRY_VENMO_APP_SWITCH
+                }).flush();
+            location.replace(window.document.referrer);
+        } else {
+            sendToParent(MESSAGE.DETECT_WEB_SWITCH);
+        }
+    });
 
     return {
         destroy
