+++ src/native/lib/index.js
@@ -1,3 +1,4 @@
 /* @flow */
 
 export * from './logger';
+export * from './appDetection';
